depr: deprecate `utils.archives` in favor of `Dataset.extract()`
